Move obfuscated settings function to utils
